Consume asynchronous snippet enumeration API
Fixes #6685

Calling IVsTextManager2.GetExpansionManager can return either an IVsExpansionManager or an IExpansionManager, and we now support reading snippet info from both.

In the IVsExpansionManager case, we behave as before this change. During package load, we kick off a task to execute later on the UI thread to synchronously enumerate the snippets.

In the IExpansionManager case, we instead enumerate the snippets asynchronously on a background thread.

For Visual Studio 2015 Update 1 RTM, we will only receive an IVsExpansionManager. Starting with Update 2 (and with some intermediate builds) we will receive IExpansionManagers by default, but the platform reserves the right to return an IVsExpansionManager as a fallback, so we must continue to support both. Because IExpansionManager already existed in Update 1, we do not need to do adaptive light-up to keep this code working on Update 1.