Fix design time builds in the Roslyn solution for PublicSign projects
Because #7812 is not yet fixed, the IDE doesn't know if we set the
PublicSign flag. As a result, all design-time builds will thing we're
real-signing, which causes semantics to get all screwed up. The
workaround for now is, for design-time builds only, to pass the
DelaySign flag since that's "good enough".

Fixes GitHub issue #7790.