Use infinite timeout when connecting to server
This suite was flaky because it used the standard timeouts when
communicating with the server.  Under high load the server could not
process requests in the standard timeout, the compilation fell back to a
local compilation and the test failed because it expected all
compilations to run on the server.

The code now uses an in process compilation which allowed us to specify
an infinite timeout when communicating with the server.  That will
alleviate the stress condition here.

closes #7618