Remove CSharpProject from the CSharpProjectShim* inheritence hierarchy
As best I can tell digging through source control history, this type
was some leftover remnants from a refactoring when we split out the
project system shims into "parts that talk to the project system" and
"parts that directly talk to the workspace" when we originally
implemented Code Lens. The code today resulted in having a non-abstract
class that had exactly one derived type, was never created directly
anywhere, and had bunch of methods with suspect implementations that
were never called since they were overridden in other places. This
simply deletes the intermediate type from the hierarchy and moves
stuff to the derived type.