Rename UpdateAnalyzerRules to something more appropriate
It really had nothing to do with analyze rules, but was just a generic
mechanism for the base type to signal to derived types it needed a
refresh of the compile options. Since having a single entry point for
that is nice anyways, I rename it and promote it. I also make it
abstract because it was very unclear why the derived types were
calling the base type and what the contract might be for that.