VB: Don't show "Each" keyword in completion list in nearly complete For loop
Fixes #4946

This is a tweak to ensure that the completion list does not display the "Each" keyword in a For loop that is already recognizable as a For loop (e.g. has an '=' token). For example, this case:

```VB
For | = 1 to 10
Next
```

The reason for this tweak is to avoid a conflict with the VB "For" snippet. If "Each" is recommended in this position, it causes the completion list to be displayed, which steals the TAB key from snippet navigation. VS 2013 did not show "Each" here either for exactly the same reason.