Fix speculation analyzer to correctly use converted type when testing conversions
Previously, I'd made a fix to the speculation analyzer to not remove necessary casts in the presence of Option Strict On. For example:

```VB
Option Strict On
Class C
	Sub M()
		Dim b As Byte
		b += CByte(1)
	End Sub
End Class
```

This was fixed to address #2560. However, the fix was less than ideal because it had to work around the fact that there isn't a compiler API
available to retrieve the built-in operator of an AssignmentStatement in VB (#6366). Essentially, here's how the workaround works:

When Option Strict is not Off and an assignment statement is being replaced with another assignment statement, the workaround checks the type
of the new expression on the right and verifies that it's convertible to the type of the expression on the right. For the example above, it checks
to see if '1' (System.Int32) is implicitly convertible to 'b' (System.Byte), which it isn't. So, the cast is left behind for Option Strict On.

**...Fast forward to now...**

Option Strict On
Class C
	Privat i As Short
	Sub M()
		Me.i = 0
	End Sub
End Class

It turns out that this was caused by the workaround described above. With that workaround, we check to see if the '0' (System.Int32) is convertible
to 'Me.i' (System.Int16). It isn't, so things go askew. The fix is to use the **converted type** of the expression on the right (System.Int16) and check
to see if it is convertible to the type on the left.