Code Model: Don't crash when adding element, deleting its text and then adding it again
Fixes #8423

In Code Model, we weakly cache code elements created by a FileCodeModel so long as they're alive. Elements are
removed from the cache when they are deleted using Code Models APIs (e.g. `FileCodeModel.Remove(...)`) or during
a clean up pass on idle. However, that means that it's possible to have collisions in the cache if a Code Model
consumer does something so:

1. Add an element, say a field called 'placeholder'.
2. Replace the text of the element in the buffer.
3. Add another field called 'placeholder'.

The steps above will likely result in a crash if the cache doesn't get cleaned up after the text is changed.

This PR adds code to check the cache before a new element is added. If an element with the same node key is found
in the cache and that element is "dead" (that is, it's node key can't be resolved), we remove the old element
from the cache.