Change how we format on certain commit..
..chars from completion. We used to call the formatting service from the
completion list with the span equal to that of the text that we
inserted, in all cases. This may not be the right approach in all cases
and that is manifested when we are formatting for a typed char that
happens to be the trigger character for the formatter as well. In such
cases we must find the right span to give to the formatter, or in other
words, the span the formatter would have come up with if it were
triggered first for the typed char. We do this by simply invoking the
appropriate formatting service API and giving it the typed char and its
position, which lets the formatting service to compute the appropriate
span.
We do this only when the completion's commit char is also
a trigger for formatting. In other cases, we continue taking the
existing code path of formatting by giving it a known span.

Fixes #7768, #8228