Fix race in Code Model cache
This addresses a race condition introduced in our per-FileCodeModel cache of code elements by PR #8424. In that change, we attempted to
handle the situation where a code model element is about to be inserted in the cache with the same "node key" is an element that is
already in the cache. This can happen in normal operation if a Code Model consumer were to programmatically add a variable, delete
the variable's underlying text and immediately add the same variable again. The original fix would conditionally remove the old
element from the cache if its "node key" could not be resolved in the FileCodeModel's current syntax tree. Then it would add the new element
with the same "node key" to the cache. However, this appears to not be deterministic and would sometimes result in a collision.

This PR changes the policy to **unconditionally** remove the old element from the cache if one exists before adding the new element.
Because the purpose of the cache is to return existing elements when accessing various collections (e.g. "Members"), it makes sense
that the latest element would be returned. In addition, because the "node key" referenced by the old element has become valid again, accessing
properties on the old element will continue to work.