Dont add IndentOperation for Broken BracketedArgumentList
Fixes internal bug 150343

On adding Indent Operation around a broken BracketedArgumentList syntax
breaks Formatting. This fix adds a check to skip the IndentOperation for
the broken code.

There might be other general cases which could trigger this Contract
Failure. But finding a repo with a broken code that triggers this
failure is difficult.

If in the future we see similar failures then we can try to generalize
these cases.