Normalize paths when generating pipe names
It turns out APIs are very inconsistent in whether they leave a trailing
directory separator on path names. This caused the client and server
code to generate different pipe names for the same file location. Hence
VbcsCompiler -shutdown was failing because it used the wrong pipe name.
Fix this by normalizing path names.