Change how Miscellaneous Files chooses which documents to track
Our previous heuristic is we tracked any document that had a moniker
that wasn't contained within any project in VisualStudioWorkspace.
This was implemented with a private hook between the two systems. This
blows up in the face of other parties wishing to create their own
workspaces against files in the RDT, since we stomp on them. The new
heuristic is we attach to any file that's open and doesn't already have
a workspace, and we watch for registration changes to back off if
something else has taken a file.

Fixes #8598.