Fix test baseline.
These tests were testing that replacement to var works where type is
apparent. but they also had another rule that said prefer built-in
types. so i update these tests to not use built-in types and replace
them with their framework counterparts. That way we're now clearly
testing if the fix works where type is apparent.