Fix race in MultipleSimultaneousCompiles
Several instances of the RunCompilationAsync helper methods are spawned off in this test and utilize the shared resource `Temp`.  If these end up resuming in parallel they can race on calls to `List<T>.Add` and throw an exception. Fix the issue by making the method `static`, and doing the shared data management in the non-async versions of the code.

closes #9630