Only create implicit project cache for VS Workspaces...
Previously, we were creating an implicit SimpleMRUCache for all Workspace
types, but we were only monitoring/cleaning up the cache in the VS case.
The effect of this is that we could end up holding on to lots of garbage
for Miscellaneous Files Workspaces even after the files had been closed
(as detailed in DevDiv 126519).  Instead, we'll only create the cache in
the VS case for now (that's the only place we know that we need it).  We
can add more Workspace types later (if needed).