Attempt to diagnose the flakiness in TestTagsChangesForEntireFile
I'm unable to figure out what is going wrong here. My best guess is that
a second workspace event happens before the text changes that causes
classification to refresh while still on the original snapshot.

This adds some additional diagnostics to our test to try to identify that.

Also changes CheckPoint.Release to CheckPoint.TryRelease, so that it's
possible to observe if someone else already caused the checkpoint task to
complete.

Related to #9723.