Prepare cibuild.cmd to receive /buildTimeLimit
/buildTimeLimit is the time limit, measured in minutes, for the Jenkins
job that runs the build. The Jenkins script netci.groovy passes the time
limit to cibuild.cmd.

netci.groovy does not yet pass the time limit to cibuild.cmd. We are
making this change to cibuild.cmd in all branches *before* modifying
netci.groovy. If we didn't do things in this order, we'd have to modify
cibuild.cmd in *all* branches *at the same time* we made the change to
netci.groovy. This way, we'll be able to first modify netci.groovy to
pass the new parameter without causing any harm. Then we'll be able to
go to each branch in turn, modifying cibuild.cmd and BuildAndTest.cmd to
actually make use of the new parameter.