Fix unit test deadlocks
There are runs on Jenkins where the IDE is failing the assertions of running on the foreground / background thread.  This causes an exception to be emitted and generally fails the resulting test.  In the case of this bug though the exception was causing the test to deadlock due to this particular call.

http://source.roslyn.io/#Microsoft.CodeAnalysis.EditorFeatures/Implementation/Classification/SyntacticClassificationTaggerProvider.TagComputer.cs,203

``` csharp
_notificationService.RegisterNotification(() =>
    {
        _workQueue.AssertIsForeground();
        ReportChangedSpan(snapshot.GetFullSpan());
    },
    ReportChangeDelayInMilliseconds,
    _listener.BeginAsyncOperation("ReportEntireFileChanged"),
    _reportChangeCancellationSource.Token);
```

Here the callback throws due to the failed foreground check.  The implementation of RegisterNotification wasn't properly hardened to ensure `IAsyncToken` was disposed in the face of an exception.  This caused that particular token to never complete and deadlocked the unit tests.

closes #7825