Fix object initalizer intellisense with generics
When doing accessibility checks to offer members initializable from the
calling typing, object initializer completion incorrectly checked that the
member being initialized was accessible from the type containing the
member instead of the type where the object initializer was being written.
This check almost always returned true, but was noticable in #10435 where
the type being initialized was a generic type constructed with a type not
accessible to the generic type. This resulted in the accessibility check
returning false. The fix is to simply check accessibility of the correct
type.