Add initial support for legacy IParameterKind interface in Code Model
We didn't implement the old, internal, legacy `IParameterKind` interface on `EnvDTE.CodeParameter` because much of the information provided by this interface was included in `EnvDTE80.CodeParameter2` back in Visual Studio 2005. However, it turns out that there is some old code in the V1 Windows Workflow designer that depends on `IParameterKind` being implemented. This change adds support for `IParameterKind` to `EnvDTE.CodeParameter` and implementations for two members: `GetParameterPassingMode(...)` and `SetParameterPassingMode(...)`.