Trigger Automation
There are 2 changes here
1. Move some of the helpers from test_util to tools_util since the helpers are not depended on the helper binary and can work without any dependencies
2. Introduce TriggerAutomation.bat, which triggers the automation process.

The reason we need a bat file to trigger this and not a csx or a new csx is because running the csx requires the helper binary. Hence we need to run the csx files from the binaries folder but we cannot have binaries in place if we dont fetch them in the first place. This is a cyclic problem that has to be broken with the help of one initial Trigger Batch file