Add a new API to get solution/document-specific options
As a part of this change, I'm converting OptionSet from being a sealed,
non-publically creatable type to an abstract type. I have chatted with
@jaredpar and @davkean and we have all concluded this is a safe
operation to make, assuming that anybody who was constructing the type
is being recompiled. Since the only creator is within Roslyn itself,
this is safe. I did have to update our PublicAPI.Shipped.txt baseline,
because our public API checker includes modifiers in the list and is
unaware that this particular modification is safe.

Two APIs are added; Solution.Options exists that will eventually give a
snapshotted OptionSet to ensure the same snapshot guarantees you get
from the rest of the Solution API, and Document.Options exists to get
the options that apply to a specific Document. I considered adding
Project.Options, but discovered in practice nobody had a meaningful use
for it. I had expected it could be used when generating new documents,
but since style options can be specified on a file/folder basis, this
wouldn't be safe to use in that case.

This API isn't implemented in a meaningful way yet, but allows for
features to consume the new system once it exists.