Fix #11193
When a SymbolId is created from a local symbol that comes from a
SpeculativeSemanticModel, GetInteriorSymbols crashes because it assumes
taht the LocalSymbol and its containing MethodSymbol have the same
SyntaxTree. With the change to use SymbolId in Completion, this causes
frequent crashes. Augment GetInteriorSymbols to not crash; this may cause
the completion tooltip to show incorrect type data if a local with the
same name is created multiple times in a method.